Modularize FSM coordination into analytics, strategy, operations, and persistence mixins
